dp: separate DataSpaceLayoutState
