+++ packages/legend-extension-dsl-data-space/src/stores/DataSpaceViewerState.ts
@@ -25,18 +25,10 @@ import type {
   GraphData,
   PackageableRuntime,
 } from '@finos/legend-graph';
-import {
-  getNonNullableEnry,
-  getNullableEntry,
-  getNullableFirstEntry,
-  guaranteeNonNullable,
-  isNonNullable,
-} from '@finos/legend-shared';
+import { getNullableEntry } from '@finos/legend-shared';
 import { action, computed, makeObservable, observable } from 'mobx';
 import type {
   DataSpaceAnalysisResult,
-  DataSpaceDiagramAnalysisResult,
-  DataSpaceExecutableAnalysisResult,
   DataSpaceExecutionContextAnalysisResult,
 } from '../graph-manager/action/analytics/DataSpaceAnalysis.js';
 import {
@@ -46,290 +38,14 @@ import {
 import { DataSpaceViewerDataAccessState } from './DataSpaceViewerDataAccessState.js';
 import { DataSpaceViewerModelsDocumentationState } from './DataSpaceModelsDocumentationState.js';
 import { DataSpaceViewerDiagramViewerState } from './DataSpaceViewerDiagramViewerState.js';
-
-export enum DATA_SPACE_VIEWER_ACTIVITY_MODE {
-  DESCRIPTION = 'description',
-  DIAGRAM_VIEWER = 'diagram-viewer',
-  MODELS_DOCUMENTATION = 'models-documentation',
-  QUICK_START = 'quick-start',
-  EXECUTION_CONTEXT = 'execution-context',
-  DATA_ACCESS = 'data-access',
-
-  DATA_STORES = 'data-stores', // TODO: with test-data, also let user call TDS query on top of these
-  DATA_AVAILABILITY = 'data-availability',
-  DATA_READINESS = 'data-readiness',
-  DATA_COST = 'data-cost',
-  DATA_GOVERNANCE = 'data-governance',
-  INFO = 'info', // TODO: test coverage? (or maybe this should be done in elements/diagrams/data-quality section)
-  SUPPORT = 'support',
-}
-
-const generateAnchorChunk = (text: string): string =>
-  encodeURIComponent(
-    text
-      .trim()
-      .toLowerCase() // anchor is case-insensitive
-      .replace(/\s+/gu, '-'), // spaces will be replaced by hyphens
-  );
-export const generateAnchorForActivity = (activity: string): string =>
-  generateAnchorChunk(activity);
-export const extractActivityFromAnchor = (anchor: string): string =>
-  decodeURIComponent(anchor);
-export const generateAnchorForQuickStart = (
-  quickStart: DataSpaceExecutableAnalysisResult,
-): string =>
-  [
-    DATA_SPACE_VIEWER_ACTIVITY_MODE.QUICK_START,
-    generateAnchorChunk(quickStart.title),
-  ].join(NAVIGATION_ZONE_SEPARATOR);
-export const generateAnchorForDiagram = (
-  diagram: DataSpaceDiagramAnalysisResult,
-): string =>
-  [
-    DATA_SPACE_VIEWER_ACTIVITY_MODE.DIAGRAM_VIEWER,
-    generateAnchorChunk(diagram.title),
-  ].join(NAVIGATION_ZONE_SEPARATOR);
-
-export const DATA_SPACE_WIKI_PAGE_SECTIONS = [
-  DATA_SPACE_VIEWER_ACTIVITY_MODE.DESCRIPTION,
-  DATA_SPACE_VIEWER_ACTIVITY_MODE.DIAGRAM_VIEWER,
-  DATA_SPACE_VIEWER_ACTIVITY_MODE.MODELS_DOCUMENTATION,
-  DATA_SPACE_VIEWER_ACTIVITY_MODE.QUICK_START,
-  DATA_SPACE_VIEWER_ACTIVITY_MODE.DATA_ACCESS,
-];
-
-const DATA_SPACE_WIKI_PAGE_ANCHORS = DATA_SPACE_WIKI_PAGE_SECTIONS.map(
-  (activity) => generateAnchorForActivity(activity),
-);
-
-type DataSpacePageNavigationCommand = {
-  anchor: string;
-};
-
-class DataSpaceLayoutState {
-  readonly dataSpaceViewerState: DataSpaceViewerState;
-
-  currentNavigationZone = '';
-  isExpandedModeEnabled = false;
-
-  frame?: HTMLElement | undefined;
-  header?: HTMLElement | undefined;
-  isTopScrollerVisible = false;
-
-  private wikiPageAnchorIndex = new Map<string, HTMLElement>();
-  wikiPageNavigationCommand?: DataSpacePageNavigationCommand | undefined;
-  private wikiPageVisibleAnchors: string[] = [];
-  private wikiPageScrollIntersectionObserver?: IntersectionObserver | undefined;
-
-  constructor(dataSpaceViewerState: DataSpaceViewerState) {
-    makeObservable<
-      DataSpaceLayoutState,
-      | 'wikiPageAnchorIndex'
-      | 'wikiPageVisibleAnchors'
-      | 'updatePageVisibleAnchors'
-    >(this, {
-      currentNavigationZone: observable,
-      isExpandedModeEnabled: observable,
-      isTopScrollerVisible: observable,
-      wikiPageAnchorIndex: observable,
-      wikiPageVisibleAnchors: observable,
-      frame: observable.ref,
-      wikiPageNavigationCommand: observable.ref,
-      isWikiPageFullyRendered: computed,
-      registerWikiPageScrollObserver: action,
-      setCurrentNavigationZone: action,
-      enableExpandedMode: action,
-      setFrame: action,
-      setTopScrollerVisible: action,
-      setWikiPageAnchor: action,
-      unsetWikiPageAnchor: action,
-      setWikiPageAnchorToNavigate: action,
-      updatePageVisibleAnchors: action,
-    });
-
-    this.dataSpaceViewerState = dataSpaceViewerState;
-  }
-
-  setCurrentNavigationZone(val: string): void {
-    this.currentNavigationZone = val;
-  }
-
-  get isWikiPageFullyRendered(): boolean {
-    return (
-      Boolean(this.frame) &&
-      DATA_SPACE_WIKI_PAGE_SECTIONS.includes(
-        this.dataSpaceViewerState.currentActivity,
-      ) &&
-      DATA_SPACE_WIKI_PAGE_ANCHORS.every((anchor) =>
-        this.wikiPageAnchorIndex.has(anchor),
-      ) &&
-      Array.from(this.wikiPageAnchorIndex.values()).every(isNonNullable)
-    );
-  }
-
-  registerWikiPageScrollObserver(): void {
-    if (this.frame && this.isWikiPageFullyRendered) {
-      const wikiPageIntersectionObserver = new IntersectionObserver(
-        (entries, observer) => {
-          const anchorsWithVisibilityChanged = entries
-            .map((entry) => {
-              for (const [key, element] of this.wikiPageAnchorIndex.entries()) {
-                if (element === entry.target) {
-                  return { key, isIntersecting: entry.isIntersecting };
-                }
-              }
-              return undefined;
-            })
-            .filter(isNonNullable);
-          anchorsWithVisibilityChanged.forEach((entry) => {
-            this.updatePageVisibleAnchors(entry.key, entry.isIntersecting);
-          });
-          // NOTE: sync scroll with menu/address is quite a delicate piece of work
-          // as it interferes with programatic scroll operations we do elsewhere.
-          // This is particularly bad when we do a programatic `smooth` scroll, which
-          // mimic user scrolling behavior and would tangle up with this observer
-          // Since currently, there's no good mechanism to detect scroll end event, and as such,
-          // there is no good way to temporarily disable this logic while doing the programmatic
-          // smooth scroll as such, we avoid supporting programatic smooth scrolling for now
-          // See https://github.com/w3c/csswg-drafts/issues/3744
-          // See https://developer.mozilla.org/en-US/docs/Web/API/Document/scrollend_event
-          if (
-            // if current navigation zone is not set, do not update zone
-            this.currentNavigationZone === '' ||
-            // if there is no visible anchors, do not update zone
-            !this.wikiPageVisibleAnchors.length ||
-            // if some of the current visible anchors match or is parent section of the current
-            // navigation zone, do not update zone
-            this.wikiPageVisibleAnchors.some(
-              (visibleAnchor) =>
-                this.currentNavigationZone === visibleAnchor ||
-                this.currentNavigationZone.startsWith(
-                  `${visibleAnchor}${NAVIGATION_ZONE_SEPARATOR}`,
-                ),
-            )
-          ) {
-            return;
-          }
-          const anchor = getNonNullableEnry(this.wikiPageVisibleAnchors, 0);
-          this.dataSpaceViewerState.syncZoneWithNavigation(anchor);
-          const anchorChunks = anchor.split(NAVIGATION_ZONE_SEPARATOR);
-          const activity = getNullableFirstEntry(anchorChunks);
-          if (activity) {
-            this.dataSpaceViewerState.setCurrentActivity(
-              extractActivityFromAnchor(
-                activity,
-              ) as DATA_SPACE_VIEWER_ACTIVITY_MODE,
-            );
-          }
-        },
-        {
-          root: this.frame,
-          threshold: 0.5,
-        },
-      );
-      Array.from(this.wikiPageAnchorIndex.values()).forEach((el) =>
-        wikiPageIntersectionObserver.observe(el),
-      );
-      this.wikiPageScrollIntersectionObserver = wikiPageIntersectionObserver;
-    }
-  }
-
-  unregisterWikiPageScrollObserver(): void {
-    this.wikiPageScrollIntersectionObserver?.disconnect();
-    this.wikiPageScrollIntersectionObserver = undefined;
-    this.wikiPageVisibleAnchors = [];
-  }
-
-  private updatePageVisibleAnchors(
-    changedAnchor: string,
-    isIntersecting: boolean,
-  ): void {
-    if (isIntersecting) {
-      const anchors = this.wikiPageVisibleAnchors.filter(
-        (anchor) => changedAnchor !== anchor,
-      );
-      // NOTE: the newly visible anchors should be the furthest one in
-      // the direction of scroll
-      anchors.push(changedAnchor);
-      this.wikiPageVisibleAnchors = anchors;
-    } else {
-      this.wikiPageVisibleAnchors = this.wikiPageVisibleAnchors.filter(
-        (anchor) => changedAnchor !== anchor,
-      );
-    }
-  }
-
-  enableExpandedMode(val: boolean): void {
-    this.isExpandedModeEnabled = val;
-  }
-
-  setFrame(val: HTMLElement | undefined): void {
-    this.frame = val;
-  }
-
-  setTopScrollerVisible(val: boolean): void {
-    this.isTopScrollerVisible = val;
-  }
-
-  setWikiPageAnchor(anchorKey: string, element: HTMLElement): void {
-    // do not allow overriding existing anchor
-    if (!this.wikiPageAnchorIndex.has(anchorKey)) {
-      this.wikiPageAnchorIndex.set(anchorKey, element);
-    }
-  }
-
-  unsetWikiPageAnchor(anchorKey: string): void {
-    this.wikiPageAnchorIndex.delete(anchorKey);
-  }
-
-  setWikiPageAnchorToNavigate(
-    val: DataSpacePageNavigationCommand | undefined,
-  ): void {
-    this.wikiPageNavigationCommand = val;
-  }
-
-  navigateWikiPageAnchor(): void {
-    if (
-      this.frame &&
-      this.wikiPageNavigationCommand &&
-      this.isWikiPageFullyRendered
-    ) {
-      const anchor = this.wikiPageNavigationCommand.anchor;
-      const matchingWikiPageSection = this.wikiPageAnchorIndex.get(anchor);
-      const anchorChunks = anchor.split(NAVIGATION_ZONE_SEPARATOR);
-      if (matchingWikiPageSection) {
-        this.frame.scrollTop =
-          matchingWikiPageSection.offsetTop -
-          (this.header?.getBoundingClientRect().height ?? 0);
-      } else if (
-        getNullableFirstEntry(anchorChunks) ===
-        generateAnchorForActivity(
-          DATA_SPACE_VIEWER_ACTIVITY_MODE.DIAGRAM_VIEWER,
-        )
-      ) {
-        this.frame.scrollTop =
-          guaranteeNonNullable(
-            this.wikiPageAnchorIndex.get(
-              generateAnchorForActivity(
-                DATA_SPACE_VIEWER_ACTIVITY_MODE.DIAGRAM_VIEWER,
-              ),
-            ),
-          ).offsetTop - (this.header?.getBoundingClientRect().height ?? 0);
-        const matchingDiagram =
-          this.dataSpaceViewerState.dataSpaceAnalysisResult.diagrams.find(
-            (diagram) => generateAnchorForDiagram(diagram) === anchor,
-          );
-        if (matchingDiagram) {
-          this.dataSpaceViewerState.diagramViewerState.setCurrentDiagram(
-            matchingDiagram,
-          );
-        }
-      }
-
-      this.setWikiPageAnchorToNavigate(undefined);
-    }
-  }
-}
+import {
+  DATA_SPACE_WIKI_PAGE_SECTIONS,
+  DataSpaceLayoutState,
+} from './DataSpaceLayoutState.js';
+import {
+  DATA_SPACE_VIEWER_ACTIVITY_MODE,
+  generateAnchorForActivity,
+} from './DataSpaceViewerNavigation.js';
 
 export class DataSpaceViewerState {
   readonly applicationStore: GenericLegendApplicationStore;
